requirements.txt
flask
pymysql
flask-sqlalchemy
flask-cors
GitPython
gunicorn
